requirements.txt
fastapi
uvicorn
pydantic[email]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-dotenv
requests
Faker
matplotlib
PyJWT
gunicorn
python-multipart
apscheduler
pytz
colorlog
b2sdk
blurhash-python 
asyncpg
brevo_python 
redis
imagekitio
face_recognition
git+https://github.com/ageitgey/face_recognition_models.git